CS506_Assignment10/requirements.txt
Flask
pandas
ipython
Pillow
torch
open_clip_torch
scikit-learn
numpy
